Add gdrive dependency and refactor navigation components in frontend
