+++ backend/environment.yml
@@ -12,4 +12,5 @@ dependencies:
   - flask-sqlalchemy
   - flask-cors
   - waitress
+  - gdrive
 prefix: /opt/miniconda3/envs/snapish
