--- frontend/src/components/layout/Footer.vue
@@ -1,14 +1,0 @@
-<template>
-  <!-- 푸터 영역 시작 -->
-  <footer class="bg-gray-100 py-4 text-center">
-    <!-- 저작권 정보 표시 -->
-    <p>&copy; 2024 SNAPish. All rights reserved.</p>
-  </footer>
-  <!-- 푸터 영역 끝 -->
-</template>
-
-<script>
-export default {
-  name: 'Footer' // 컴포넌트 이름 정의
-}
-</script>
