+++ frontend/src/views/FishWarning.vue
@@ -83,27 +83,14 @@
       </main>
 
       <!-- 하단 네비게이션 바 -->
-      <nav class="fixed bottom-0 left-0 right-0 bg-white border-t px-6 py-2 max-w-md mx-auto">
-        <div class="flex justify-between items-center">
-          <button class="flex flex-col items-center p-2">
-            <HomeIcon class="w-6 h-6" />
-            <span class="text-xs mt-1">홈</span>
-          </button>
-          <button class="flex flex-col items-center p-2">
-            <CameraIcon class="w-6 h-6" />
-            <span class="text-xs mt-1">촬영</span>
-          </button>
-          <button class="flex flex-col items-center p-2">
-            <UserIcon class="w-6 h-6" />
-            <span class="text-xs mt-1">프로필</span>
-          </button>
-        </div>
-      </nav>
+      <BottomNavigation @toggleCameraActions="showCameraActions = true" />
     </div>
   </div>
 </template>
 
 <script setup>
+import BottomNavigation from '../components/layout/BottomNavigation.vue';
+
 // Lucide 아이콘 컴포넌트 불러오기
 import {
   BellIcon,         // 알림 아이콘
@@ -112,9 +99,6 @@ import {
   AlertTriangleIcon,// 경고 아이콘
   InfoIcon,         // 정보 아이콘
   Share2Icon,       // 공유 아이콘
-  HomeIcon,         // 홈 아이콘
-  CameraIcon,       // 카메라 아이콘
-  UserIcon          // 사용자 아이콘
 } from 'lucide-vue-next'
 </script>
